Split template definition
