+++ template/light.go
@@ -1,0 +1,4 @@
+package template
+
+type Light struct {
+}
