+++ template/template.go
@@ -18,19 +18,9 @@ type Template struct {
 
 	// Aircon template
 	Aircon *Aircon `json:"aircon,omitempty"`
-}
 
-type Aircon struct {
-	Operation *Action                `json:"operation"`
-	Modes     map[string]*AirconMode `json:"modes"`
-}
-
-type AirconMode struct {
-	Temp           *Action `json:"temp,omitempty"`
-	Humid          *Action `json:"humid,omitempty"`
-	Fan            *Action `json:"fan,omitempty"`
-	HorizontalVane *Action `json:"horizontal_vane,omitempty"`
-	VerticalVane   *Action `json:"vertical_vane,omitempty"`
+	// Light template
+	Light *Light `json:"light,omitempty"`
 }
 
 type Action struct {
